requirements.txt
# pip libraries required
art
climage
pillow
termcolor
tabulate